Move canDeserialize logic to base classes
