+++ src/conditions/never.ts
@@ -1,10 +1,5 @@
 import { BaseCondition } from './base';
-import { Serialized } from '../interfaces/serialized';
 import { NothingSelector } from '../selector/input/nothing';
-
-interface NeverSerialized extends Serialized {
-  class: 'Never';
-}
 
 export class Never extends BaseCondition {
   constructor() {
@@ -13,9 +8,6 @@ export class Never extends BaseCondition {
 
   static deserialize() {
     return new Never();
-  }
-  static canDeserialize(obj: Serialized): obj is NeverSerialized {
-    return obj.class === 'Never';
   }
 
   serialize() {
