+++ src/deserializer.ts
@@ -1,7 +1,7 @@
-import { Serialized } from '../interfaces/serialized';
-import { Deserializable } from '../interfaces/deserializable';
-import { DeserializationError } from './errors/deserialization-error';
-import { Serializable } from '../interfaces/serializable';
+import { Serialized } from './interfaces/serialized';
+import { Deserializable } from './interfaces/deserializable';
+import { DeserializationError } from './lib/errors/deserialization-error';
+import { Serializable } from './interfaces/serializable';
 
 export abstract class Deserializer<T extends Serializable<Serialized>> {
   protected abstract deserializers: Deserializable<T>[];
