+++ src/selector/input/base.ts
@@ -3,6 +3,9 @@ import { Serialized } from '../../interfaces/serialized';
 import { Serializable } from '../../interfaces/serializable';
 
 export abstract class BaseInputSelector implements Serializable<Serialized> {
+  static canDeserialize(obj: Serialized): boolean {
+    return obj.class === this.name;
+  }
   abstract apply(input: Input): any;
   abstract serialize(): Serialized;
 }
